Start changing lsp code for handling multiple instances
